refactor(renderer): split translate page feature
