+++ src/renderer/src/features/translate/components/styles.ts
@@ -1,0 +1,8 @@
+export const btnBase =
+  'inline-flex items-center gap-1.5 h-[30px] px-3 rounded-md border border-neutral-700 bg-[#131518] text-neutral-200 text-xs font-medium cursor-pointer transition-all hover:bg-neutral-800 hover:border-neutral-600 active:translate-y-px select-none whitespace-nowrap'
+
+export const btnGhostIcon =
+  'inline-flex items-center justify-center w-[30px] h-[30px] rounded-md bg-transparent border-0 text-neutral-400 cursor-pointer transition-all hover:bg-neutral-800 hover:text-neutral-200 select-none'
+
+export const btnPrimary =
+  'inline-flex items-center gap-1.5 h-[30px] px-3 rounded-md border border-amber-500 bg-amber-500 text-neutral-950 text-xs font-semibold cursor-pointer transition-all hover:bg-amber-400 hover:border-amber-400 active:translate-y-px select-none whitespace-nowrap'
