+++ src/renderer/src/features/translate/types.ts
@@ -1,0 +1,4 @@
+import type { useTranslationSession } from '@/context/TranslationSession'
+
+export type ExportFormat = 'xml' | 'pak' | 'zip'
+export type TranslationSession = ReturnType<typeof useTranslationSession>
